When the player is available for Calls, the CallQueue will dispatch a call to the player

+++ MonoDragons.Core/Common/RNG.cs
@@ -11,7 +11,7 @@ namespace MonoDragons.Core.Common
             return Instance.Next(max);
         }
 
-        public static float Int(int min, int max)
+        public static int Int(int min, int max)
         {
             return Instance.Next(min, max);
         }
